+++ canvas-component-tools/CHANGELOG.md
@@ -4,7 +4,8 @@ Canvas Component Tools uses [semantic versioning](https://semver.org/).
 
 ## 1.1.0 - 2022-31-01
 
-* Added limited interactivity template option
+* Added limited interactivity to template options
+* Added documentation links to footer as choices are made
 * Updated template titles to not use 404 title to avoid confusion
 * Updated command to output version of the tool instead of Canvas
 * Updated dependencies